Parameters for shop modules
Summary:
Add 3 new fields for a shop module to param it :
  - logout_post_purchase (boolean), if True the user is going to be logout after a sale,
  - limit_purchase (Decimal), limit the maximum amount for a sale,
  - delay_post_purchase (Integer), number of seconds before automatically leaving the resume view of a sale.

All of them can be changed in the view "workboard" and "config" of a module (self or operator).

Change the original success_url of a sale_interface to be the sale_interface itself (if logout_after_purchase is False).

Test Plan: .

Reviewers: pastys

Reviewed By: pastys

Differential Revision: http://phabricator.iresam.org/D840